Tiles (cpp/h) part 3

+++ src/RaylibWrapper.cpp
@@ -2,7 +2,7 @@
 // Created by lehen on 12/11/2023.
 //
 
-#include "../headers/raylibWrapper.h"
+#include "../headers/RaylibWrapper.h"
 #include <cassert>
 
 namespace GameEngine {
@@ -34,10 +34,10 @@ namespace GameEngine {
 
     bool CheckCollisionPointRec(const Vec2<int> point, const Vec2<int> position, const Vec2<int> size) {
         Vector2 targetPoint{static_cast<float>(point.GetX()), static_cast<float>(point.GetY())};
-        Rectangle rec{position.GetX(),
-                      position.GetY(),
-                      size.GetX(),
-                      size.GetY()};
+        Rectangle rec{static_cast<float>(position.GetX()),
+                      static_cast<float>(position.GetY()),
+                      static_cast<float>(size.GetX()),
+                      static_cast<float>(size.GetY())};
         return CheckCollisionPointRec(targetPoint, rec);
     }
